refactor: rename compliance trigger to onSecurityScanCompleted
Rework the package-processed trigger into a security/vulnerability trigger:

- Rename cloudsmith.onComplianceCheckCompleted -> cloudsmith.onSecurityScanCompleted,
  backed by the package.security_scanned event (fires when the scan completes,
  so the vulnerability data is settled rather than pending at sync time).
- Read the vulnerability summary (has_vulnerabilities, max_severity,
  num_vulnerabilities) directly from the webhook payload's
  context.vulnerability_scan_results — no extra API call. Drop the best-effort
  GetPackageVulnerabilities client method.
- Emit cloudsmith.package.securityScanned with package identity + scan results.
- Extract shared webhook helpers (signature verification, package parsing,
  RepositoryRef) into webhook_trigger.go.
- Update frontend renderer, tests, example data, and regenerate docs.
