+++ pkg/integrations/cloudsmith/cloudsmith.go
@@ -75,7 +75,7 @@ func (c *Cloudsmith) Actions() []core.Action {
 
 func (c *Cloudsmith) Triggers() []core.Trigger {
 	return []core.Trigger{
-		&OnComplianceCheckCompleted{},
+		&OnSecurityScanCompleted{},
 		&OnPackageCreated{},
 	}
 }
